Halfway docstringified, pushing to test run refactored location
